requirements.txt
requests
python-dotenv
Pillow
beautifulsoup4